working on some problems such as
There should be a separate subsection inside the conclusion section called 'Limitations' where you would have to discuss the limitations of this dataset which might have adversely affected your analysis. Examples would be null or missing values, whether these samples are an effective representation of the population or not or maybe that you could dive deeper into your analysis with additional specific information.

The conclusions and limitations have the following structure:

Not every analysis is followed by an explanation.

Visualizations made in the project depict the data in an appropriate manner that allows plots to be readily interpreted.