fixing some errors
Please try to explore more plots in the multivariate exploration section by asking different questions from different perspectives.

A very nice job including various visualizations. Each of them is associated with comments that accurately depict their purpose but currently the code is also displayed in the presentation. In real life, when you present something, it should be as Layman as possible and hence code used to generate the plot needs to be removed. Please refer to this answer on how to generate the slides notebook without code.

Visualization presented clearly depict the data and represent the questions posed but the visualizations are missing either title or x and y axes or both. Please try to polish the visualizations for the presentation deck.